support @base/debug module
