--- DebugInfos.d.ts
@@ -1,15 +1,0 @@
-
-/**
- * This is a declaration file for 'DebugInfos.json'
- * which is generated from 'EngineErrorMap.md'.
- * You may run cli command
- * `gulp build-debug-infos` or `npx gulp build-debug-infos` to generate the json file.
- */
-
-/**
- * Engine error maps.
- * The keys are error codes and their values are format strings.
- */
-declare const $: Record<number, string>;
-
-export default $;
