+++ package.json
@@ -21,9 +21,9 @@
     "server": "http-server . -p 8002 -o playground/index.html",
     "postinstall": "npm run build",
     "build": "turbo build && npm run build-root",
-    "build-root": "gulp build-debug-infos && gulp build-declarations && npm run bundle-adapter && npm run fetch-ral",
-    "build:dev": "gulp build-h5-source",
-    "build:min": "gulp build-h5-minified",
+    "build-root": "gulp build-declarations && npm run bundle-adapter && npm run fetch-ral",
+    "build-engine": "turbo build && gulp build-h5-minified",
+    "build-engine-dev": "turbo build && gulp build-h5-source",
     "build-api-json": "gulp build-api-json -i index.ts -o ./3d-api-docs/tempJson",
     "build-api": "npm run build-api-json && gulp build-3d-api -i index.ts -o ./3d-api-docs -j ./3d-api-docs/tempJson",
     "checkout-engine-version": "node ./scripts/checkout-engine-version.js",
